hv:Fix Implict conversion:actual to formal param
MISRAC has requirement about implict conversion: actual to formal
param. This patch is used to fix part of these violations.

1.Add a new structure seg_desc_vmcs to hold the VMCS field address of
segment selector to clean up seg_desc structure.

2.Add the definition of maximum MSI entry and the relevant judgement.

3.The violations in shell.c, logmsg.c will be fixed in other series of
patches with modification of function snprintf(), vsnprintf() and other
related usages.

v1->v2:
  *Move the definition of struct seg_desc_vmcs from instr_emul.h to
   instr_emul.c.
  *Modify the formal parameter type in function definition from uint8_t
   to char instead of using cast.
  *Drop the const declaration for char data in formal parameter.

v2->v3:
  *update the data missing conversion.
  *change type of internal parameter len to avoid casting in npklog.c.
  *change the conversion from signed char to unsigned int in
   uart16550_getc() to solve sign-extension.

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/irq.c
@@ -45,7 +45,8 @@ uint32_t alloc_irq_num(uint32_t req_irq)
 	}
 
 	if (irq != IRQ_INVALID) {
-		bitmap_set_nolock(irq & 0x3FU, irq_alloc_bitmap + (irq >> 6U));
+		bitmap_set_nolock((uint16_t)(irq & 0x3FU),
+			      irq_alloc_bitmap + (irq >> 6U));
 	}
 	spinlock_irqrestore_release(&irq_alloc_spinlock, rflags);
 	return irq;
@@ -67,7 +68,7 @@ void free_irq_num(uint32_t irq)
 	if ((irq_is_gsi(irq) == false)
 	    && (desc->vector <= VECTOR_DYNAMIC_END)) {
 		spinlock_irqsave_obtain(&irq_alloc_spinlock, &rflags);
-		bitmap_test_and_clear_nolock(irq & 0x3FU,
+		bitmap_test_and_clear_nolock((uint16_t)(irq & 0x3FU),
 					     irq_alloc_bitmap + (irq >> 6U));
 		spinlock_irqrestore_release(&irq_alloc_spinlock, rflags);
 	}
@@ -333,7 +334,8 @@ void dispatch_interrupt(struct intr_excp_ctx *ctx)
 		goto ERR;
 	}
 
-	if (bitmap_test(irq & 0x3FU, irq_alloc_bitmap + (irq >> 6U)) == 0U) {
+	if (bitmap_test((uint16_t)(irq & 0x3FU),
+		irq_alloc_bitmap + (irq >> 6U)) == 0U) {
 		/* mask irq if possible */
 		goto ERR;
 	}
@@ -403,7 +405,8 @@ void get_cpu_interrupt_info(char *str_arg, int str_max)
 
 	for (irq = 0U; irq < NR_IRQS; irq++) {
 		vector = irq_to_vector(irq);
-		if (bitmap_test(irq & 0x3FU, irq_alloc_bitmap + (irq >> 6U))
+		if (bitmap_test((uint16_t)(irq & 0x3FU),
+			irq_alloc_bitmap + (irq >> 6U))
 			&& (vector != VECTOR_INVALID)) {
 			len = snprintf(str, size, "\r\n%d\t0x%X", irq, vector);
 			size -= len;
@@ -441,7 +444,8 @@ static void init_irq_descs(void)
 
 		irq_desc_array[irq].vector = vr;
 		vector_to_irq[vr] = irq;
-		bitmap_set_nolock(irq & 0x3FU, irq_alloc_bitmap + (irq >> 6U));
+		bitmap_set_nolock((uint16_t)(irq & 0x3FU),
+			      irq_alloc_bitmap + (irq >> 6U));
 	}
 }
